chore(UI): move json parset to common [YTFRONT-4982]
