+++ packages/ui/src/ui/pages/query-tracker/Navigation/api/loadDynamicTable.ts
@@ -5,7 +5,7 @@ import Query from '../../../../utils/navigation/content/table/query';
 import {injectColumnsFromSchema} from '../../../../utils/navigation/content/table/table-ts';
 import {NavigationTableSchema} from '../../module/queryNavigation/queryNavigationSlice';
 import {UnipikaValue} from '../../../../components/Yson/StructuredYson/StructuredYsonTypes';
-import {JSONParser} from '../../module/api';
+import {JSONParser} from '../../../../common/yt-api';
 import {TypeArray} from '../../../../components/SchemaDataType/dataTypes';
 import {getClusterProxy} from '../../../../store/selectors/global';
 import {ClusterConfig} from '../../../../../shared/yt-types';
